+++ pokemon-wiki-spider/src/main/java/cn/fantasticmao/pokemon/spider/task1/MoveListSpider.java
@@ -3,8 +3,6 @@ package cn.fantasticmao.pokemon.spider.task1;
 import cn.fantasticmao.mundo.core.support.Constant;
 import cn.fantasticmao.pokemon.spider.Config;
 import cn.fantasticmao.pokemon.spider.PokemonDataSource;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.commons.lang3.ObjectUtils;
@@ -18,7 +16,7 @@ import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
-import java.util.function.Function;
+import java.util.function.BiFunction;
 import java.util.stream.Collectors;
 
 /**
@@ -81,9 +79,6 @@ public class MoveListSpider extends AbstractTask1Spider<MoveListSpider.Data> {
     }
 
     @Getter
-    @Setter
-    @Builder
-    @AllArgsConstructor
     static class Data implements AbstractTask1Spider.Data {
         private final String nameZh;
         private final String nameJa;
@@ -94,9 +89,22 @@ public class MoveListSpider extends AbstractTask1Spider<MoveListSpider.Data> {
         private final String accuracy;
         private final String pp;
         private final int generation;
+
+        public Data(String nameZh, String nameJa, String nameEn, String type, String category, String power,
+                    String accuracy, String pp, int generation) {
+            this.nameZh = nameZh;
+            this.nameJa = nameJa;
+            this.nameEn = nameEn;
+            this.type = type;
+            this.category = category;
+            this.power = power;
+            this.accuracy = accuracy;
+            this.pp = pp;
+            this.generation = generation;
+        }
     }
 
-    private static final Function<Element, MoveListSpider.Data.DataBuilder> PARSER = element -> {
+    private static final BiFunction<Element, Integer, Data> PARSER = (element, generation) -> {
         String nameZh = element.child(1).text();
         String nameJa = element.child(2).text();
         String nameEn = element.child(3).text();
@@ -105,20 +113,14 @@ public class MoveListSpider extends AbstractTask1Spider<MoveListSpider.Data> {
         String power = "—".equals(element.child(6).text()) ? null : element.child(6).text();
         String accuracy = "—".equals(element.child(7).text()) ? null : element.child(7).text();
         String pp = element.child(8).text();
-        return new Data.DataBuilder()
-            .nameZh(nameZh).nameJa(nameJa).nameEn(nameEn)
-            .type(type).category(category)
-            .power(power).accuracy(accuracy).pp(pp);
+        return new Data(nameZh, nameJa, nameEn, type, category, power, accuracy, pp, generation);
     };
 
     // 第一世代
     private List<MoveListSpider.Data> getData1(Document document) {
         return document.select(".bg-关都 > tbody > tr").stream()
             .skip(1)
-            .map(element -> PARSER.apply(element)
-                .generation(1)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 1))
             .collect(Collectors.toList());
     }
 
@@ -126,10 +128,7 @@ public class MoveListSpider extends AbstractTask1Spider<MoveListSpider.Data> {
     private List<MoveListSpider.Data> getData2(Document document) {
         return document.select(".bg-城都 > tbody > tr").stream()
             .skip(1)
-            .map(element -> PARSER.apply(element)
-                .generation(2)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 2))
             .collect(Collectors.toList());
     }
 
@@ -137,10 +136,7 @@ public class MoveListSpider extends AbstractTask1Spider<MoveListSpider.Data> {
     private List<MoveListSpider.Data> getData3(Document document) {
         return document.select(".bg-丰缘 > tbody > tr").stream()
             .skip(1)
-            .map(element -> PARSER.apply(element)
-                .generation(3)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 3))
             .collect(Collectors.toList());
     }
 
@@ -148,10 +144,7 @@ public class MoveListSpider extends AbstractTask1Spider<MoveListSpider.Data> {
     private List<MoveListSpider.Data> getData4(Document document) {
         return document.select(".bg-神奥 > tbody > tr").stream()
             .skip(1)
-            .map(element -> PARSER.apply(element)
-                .generation(4)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 4))
             .collect(Collectors.toList());
     }
 
@@ -159,10 +152,7 @@ public class MoveListSpider extends AbstractTask1Spider<MoveListSpider.Data> {
     private List<MoveListSpider.Data> getData5(Document document) {
         return document.select(".bg-合众 > tbody > tr").stream()
             .skip(1)
-            .map(element -> PARSER.apply(element)
-                .generation(5)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 5))
             .collect(Collectors.toList());
     }
 
@@ -171,10 +161,7 @@ public class MoveListSpider extends AbstractTask1Spider<MoveListSpider.Data> {
         return document.select(".bg-卡洛斯 > tbody > tr").stream()
             .filter(element -> element.child(0).children().size() == 0)
             .skip(1)
-            .map(element -> PARSER.apply(element)
-                .generation(6)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 6))
             .collect(Collectors.toList());
     }
 
@@ -183,10 +170,7 @@ public class MoveListSpider extends AbstractTask1Spider<MoveListSpider.Data> {
         return document.select(".bg-阿罗拉 > tbody > tr").stream()
             .filter(element -> element.child(0).children().size() == 0)
             .skip(1)
-            .map(element -> PARSER.apply(element)
-                .generation(7)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 7))
             .collect(Collectors.toList());
     }
 
@@ -195,10 +179,7 @@ public class MoveListSpider extends AbstractTask1Spider<MoveListSpider.Data> {
         return document.select(".bg-伽勒尔 > tbody > tr").stream()
             .filter(element -> element.child(0).children().size() == 0)
             .skip(1)
-            .map(element -> PARSER.apply(element)
-                .generation(8)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 8))
             .collect(Collectors.toList());
     }
 }