+++ pokemon-wiki-spider/src/main/java/cn/fantasticmao/pokemon/spider/task1/PokemonAbilityListSpider.java
@@ -3,10 +3,7 @@ package cn.fantasticmao.pokemon.spider.task1;
 import cn.fantasticmao.mundo.core.support.Constant;
 import cn.fantasticmao.pokemon.spider.Config;
 import cn.fantasticmao.pokemon.spider.PokemonDataSource;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 import org.apache.commons.lang3.ObjectUtils;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
@@ -17,7 +14,7 @@ import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
-import java.util.function.Function;
+import java.util.function.BiFunction;
 import java.util.stream.Collectors;
 
 /**
@@ -79,9 +76,6 @@ public class PokemonAbilityListSpider extends AbstractTask1Spider<PokemonAbility
     }
 
     @Getter
-    @Setter
-    @Builder
-    @AllArgsConstructor
     static class Data implements AbstractTask1Spider.Data {
         private final int index;
         private final String nameZh;
@@ -91,9 +85,21 @@ public class PokemonAbilityListSpider extends AbstractTask1Spider<PokemonAbility
         private final String ability2;
         private final String abilityHide;
         private final int generation;
+
+        public Data(int index, String nameZh, String type1, String type2, String ability1, String ability2,
+                    String abilityHide, int generation) {
+            this.index = index;
+            this.nameZh = nameZh;
+            this.type1 = type1;
+            this.type2 = type2;
+            this.ability1 = ability1;
+            this.ability2 = ability2;
+            this.abilityHide = abilityHide;
+            this.generation = generation;
+        }
     }
 
-    private static final Function<Element, PokemonAbilityListSpider.Data.DataBuilder> PARSER = element -> {
+    private static final BiFunction<Element, Integer, Data> PARSER = (element, generation) -> {
         int index = Integer.parseInt(element.child(0).text());
         String nameZh = element.child(2).text();
         String type1 = element.child(3).text();
@@ -101,21 +107,14 @@ public class PokemonAbilityListSpider extends AbstractTask1Spider<PokemonAbility
         String ability1 = element.child(5).text();
         String ability2 = element.child(6).hasClass("hide") ? null : element.child(6).text();
         String abilityHide = element.child(7).children().size() == 0 ? null : element.child(7).text();
-        return new Data.DataBuilder()
-            .index(index).nameZh(nameZh)
-            .type1(type1).type2(type2)
-            .ability1(ability1).ability2(ability2)
-            .abilityHide(abilityHide);
+        return new Data(index, nameZh, type1, type2, ability1, ability2, abilityHide, generation);
     };
 
     // 关都地区
     private List<PokemonAbilityListSpider.Data> getDataList1(Document document) {
         return document.select(".bg-关都 > tbody > tr").parallelStream()
             .filter(element -> element.hasClass("bgwhite"))
-            .map(element -> PARSER.apply(element)
-                .generation(1)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 1))
             .collect(Collectors.toList());
     }
 
@@ -123,10 +122,7 @@ public class PokemonAbilityListSpider extends AbstractTask1Spider<PokemonAbility
     private List<PokemonAbilityListSpider.Data> getDataList2(Document document) {
         return document.select(".bg-城都 > tbody > tr").parallelStream()
             .filter(element -> element.hasClass("bgwhite"))
-            .map(element -> PARSER.apply(element)
-                .generation(2)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 2))
             .collect(Collectors.toList());
     }
 
@@ -134,10 +130,7 @@ public class PokemonAbilityListSpider extends AbstractTask1Spider<PokemonAbility
     private List<PokemonAbilityListSpider.Data> getDataList3(Document document) {
         return document.select(".bg-丰缘 > tbody > tr").parallelStream()
             .filter(element -> element.hasClass("bgwhite"))
-            .map(element -> PARSER.apply(element)
-                .generation(3)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 3))
             .collect(Collectors.toList());
     }
 
@@ -145,10 +138,7 @@ public class PokemonAbilityListSpider extends AbstractTask1Spider<PokemonAbility
     private List<PokemonAbilityListSpider.Data> getDataList4(Document document) {
         return document.select(".bg-神奥 > tbody > tr").parallelStream()
             .filter(element -> element.hasClass("bgwhite"))
-            .map(element -> PARSER.apply(element)
-                .generation(4)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 4))
             .collect(Collectors.toList());
     }
 
@@ -156,10 +146,7 @@ public class PokemonAbilityListSpider extends AbstractTask1Spider<PokemonAbility
     private List<PokemonAbilityListSpider.Data> getDataList5(Document document) {
         return document.select(".bg-合众 > tbody > tr").parallelStream()
             .filter(element -> element.hasClass("bgwhite"))
-            .map(element -> PARSER.apply(element)
-                .generation(5)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 5))
             .collect(Collectors.toList());
     }
 
@@ -167,10 +154,7 @@ public class PokemonAbilityListSpider extends AbstractTask1Spider<PokemonAbility
     private List<PokemonAbilityListSpider.Data> getDataList6(Document document) {
         return document.select(".bg-卡洛斯 > tbody > tr").parallelStream()
             .filter(element -> element.hasClass("bgwhite"))
-            .map(element -> PARSER.apply(element)
-                .generation(6)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 6))
             .collect(Collectors.toList());
     }
 
@@ -178,10 +162,7 @@ public class PokemonAbilityListSpider extends AbstractTask1Spider<PokemonAbility
     private List<PokemonAbilityListSpider.Data> getDataList7(Document document) {
         return document.select(".bg-阿羅拉 > tbody > tr").parallelStream()
             .filter(element -> element.hasClass("bgwhite"))
-            .map(element -> PARSER.apply(element)
-                .generation(7)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 7))
             .collect(Collectors.toList());
     }
 
@@ -189,10 +170,7 @@ public class PokemonAbilityListSpider extends AbstractTask1Spider<PokemonAbility
     private List<PokemonAbilityListSpider.Data> getDataList8(Document document) {
         return document.select(".bg-伽勒尔 > tbody > tr").parallelStream()
             .filter(element -> element.hasClass("bgwhite"))
-            .map(element -> PARSER.apply(element)
-                .generation(8)
-                .build()
-            )
+            .map(element -> PARSER.apply(element, 8))
             .collect(Collectors.toList());
     }
 }