+++ pokemon-wiki-web/src/main/java/cn/fantasticmao/pokemon/web/service/impl/MoveServiceImpl.java
@@ -60,7 +60,7 @@ public class MoveServiceImpl implements MoveService {
         if (page < 0 || size < 1) {
             return Collections.emptyList();
         }
-        List<Move> moveList = moveRepository.find(PageUtil.offset(page, size), PageUtil.size(size));
+        List<Move> moveList = moveRepository.find(PageUtil.offset(page, size), PageUtil.limit(size));
         return moveList.stream()
             .map(move -> MoveBean.ofDomain(move, null))
             .collect(Collectors.toList());